pyproject.toml
[project]
name = "fastapi-production-boilerplate"
version = "0.1.0"
description = "A SOLID FastAPI backend production boilerplate"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Tosh Gitonga", email = "gitongachas@gmail.com" }
]

dependencies = [
    "fastapi[standard]>=0.115.0",
    "sentry-sdk[fastapi]>=1.40.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.2.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "aiosqlite>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",

    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",

    "ipython>=8.27.0",
    "ipdb>=0.13.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # unused arguments
    "SIM", # simplifications
]
ignore = [
    "E501",  # line length handled by formatter
    "B008",  # function calls in defaults
    "B904",  # raise-from enforcement
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["ARG"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=html",
    "--cov-report=term-missing:skip-covered",
    "--cov-fail-under=80",
]

markers = [
    "slow: slow tests",
    "unit: unit tests",
    "integration: integration tests",
    "security: security-related tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/alembic/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "ipython>=9.8.0",
]
